feat: secure api routes

+++ api/src/Controller/CalendarController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Calendar;
 use App\Entity\Kid;
+use App\Entity\User;
 use App\Handler\CalendarHandler;
 use App\Repository\CalendarRepository;
 use App\Repository\FamilyRepository;
@@ -46,10 +47,11 @@ class CalendarController extends AbstractController
     }
 
     #[IsGranted('ROLE_NURSE', message: 'Vous ne pouvez pas faire ça')]
-    #[Route('/calendar/nurse/{nurseId}', name: 'app_calendar_nurse', methods: 'GET')]
-    public function calendarNurse(int $nurseId): JsonResponse
+    #[Route('/calendar/nurse/{nurse}', name: 'app_calendar_nurse', methods: 'GET')]
+    public function calendarNurse(User $nurse): JsonResponse
     {
-        return $this->json($this->calendarRepository->getCalendarByNurse($nurseId), Response::HTTP_OK);
+        $this->denyAccessUnlessGranted('owner', $nurse);
+        return $this->json($this->calendarRepository->getCalendarByNurse($nurse->getId()), Response::HTTP_OK);
     }
 
     /**
@@ -74,16 +76,18 @@ class CalendarController extends AbstractController
     #[Route('/calendar/{calendar}', name: 'app_calendar_delete', methods: 'DELETE')]
     public function delete(Calendar $calendar): JsonResponse
     {
+        $this->denyAccessUnlessGranted('owner', $calendar->getKid());
         $this->calendarHandler->handleDeleteCalendar($calendar);
 
         return $this->json([], Response::HTTP_NO_CONTENT);
     }
 
     #[IsGranted('ROLE_PARENT', message: 'Vous ne pouvez pas faire ça')]
-    #[Route('/calendar/family/{familyId}', name: 'app_calendar_family', methods: 'GET')]
-    public function calendarFamily(int $familyId): JsonResponse
+    #[Route('/calendar/family/{family}', name: 'app_calendar_family', methods: 'GET')]
+    public function calendarFamily(User $family): JsonResponse
     {
-        $family = $this->familyRepository->findOneBy(['parent' => $familyId]);
+        $this->denyAccessUnlessGranted('owner', $family);
+        $family = $this->familyRepository->findOneBy(['parent' => $family->getId()]);
         return $this->json($this->calendarRepository->getCalendarByFamily($family->getId()), Response::HTTP_OK);
     }
 }
